moved into a package called get_weather
