+++ src/use_case/get_weather/GetWeatherInteractor.java
@@ -1,4 +1,4 @@
-package use_case.weather;
+package use_case.get_weather;
 
 public class GetWeatherInteractor implements GetWeatherInputBoundary{
