+++ src/use_case/get_weather/GetWeatherOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.weather;
+package use_case.get_weather;
 
 public interface GetWeatherOutputBoundary {
     void prepareSuccessView(GetWeatherOutputData getWeatherOutputData);
